Add a fallback check for the 2.x backup configuration
When moving forward etcd runs a configuration transaction and renames
the original 2.x series config to `etcd.conf.2x`. This means the user
cannot reasonably roll back and expect etcd to continue to function.
Since COMMON is the melting pot of configuration I feel it makes sense
to look for the 2.x configuration and use it as a fallback mechanism
before defaulting to error.